requirements.txt
requests
djangorestframework
markdown
django-filter
drf-extensions